+++ crates/tuirealm/src/core/component.rs
@@ -2,9 +2,11 @@
 //!
 //! This module exposes the component traits
 
+use ratatui::Frame;
+
 use crate::command::{Cmd, CmdResult};
 use crate::ratatui::layout::Rect;
-use crate::{AttrValue, Attribute, Event, Frame, State};
+use crate::{AttrValue, Attribute, Event, State};
 
 /// A Mock Component represents a component which defines all the properties and states it can handle and represent
 /// and the way it should be rendered. It must also define how to behave in case of a `Cmd` (command).